sphere/requirements.txt
Django==1.10.5
python-telegram-bot==5.3.1
gunicorn==19.7.1
mysqlclient==1.3.9
